Split up impurity testing
